feat: added automigration with liquibase

+++ spring/src/main/java/com/lukaszfabia/main/dto/ProductDTO.java
@@ -23,7 +23,7 @@ public record ProductDTO(
     }
 
     public ProductDTO(Product product) {
-        this(product.getId(), product.getName(), product.getWeight(), product.getPrice(),
+        this(product.getId(), product.getName(), product.getWeight().doubleValue(), product.getPrice().doubleValue(),
                 product.getCategory());
     }
